可视化采集

+++ qk-reptile/qk-data-reptile/src/main/java/com/qk/dm/reptile/service/RptDimensionInfoService.java
@@ -12,7 +12,7 @@ public interface RptDimensionInfoService {
 
   void addRptDir(RptDimensionInfoDTO rptDimensionInfoDTO);
 
-  void deleteRptDir(Integer id);
+  void deleteRptDir(Long id);
 
   void updateRptDir(RptDimensionInfoDTO rptDimensionInfoDTO);
 }
